requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
langchain
langchain-core
langchain-community
faiss-cpu
langchain_google_genai